[Jagdish] add project with some editing

+++ controllers/reviews.js
@@ -7,10 +7,6 @@ module.exports.createReview=async(req,res)=>{
     req.flash("error", "Listing not found!");
     return res.redirect("/listings");
   }
-
-    // let newReview=new Review(req.body.review);
-    // newReview.author=req.user._id;
-    // listing.reviews.push(newReview);
     const { rating, comment } = req.body.review;
 
     const newReview = new Review({
@@ -30,6 +26,23 @@ module.exports.createReview=async(req,res)=>{
 
 module.exports.deleteReview=async(req,res)=>{
     let { id, reviewId } = req.params;
+    const review = await Review.findById(reviewId);
+    
+    if (!review) {
+        req.flash("error", "Review not found!");
+        return res.redirect(`/listings/${id}`);
+    }
+    
+    if (!req.user) {
+        req.flash("error", "You must be logged in to delete reviews!");
+        return res.redirect(`/listings/${id}`);
+    }
+    
+    if (!review.author || !review.author.equals(req.user._id)) {
+        req.flash("error", "You are not authorized to delete this review!");
+        return res.redirect(`/listings/${id}`);
+    }
+    
     await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
     await Review.findByIdAndDelete(reviewId);
      req.flash("success","Review Deleted!");
